refactor: move shared code to dedicated directory
