+++ src/modules/analytics/route.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import { analyticsController } from '@/modules/analytics/controller/controller';
-import asyncHandler from '@/utils/asyncHandler';
+import asyncHandler from '@/shared/utils/asyncHandler';
 
 const router = Router();
 router.get('/drr-by-date', asyncHandler(analyticsController.getDrrByDate));
