+++ src/shared/types/AppError.ts
